+++ ProjetoMVC/Models/Dtos/RegistroVendasDto.cs
@@ -4,11 +4,11 @@ namespace ProjetoMVC.Models.Dtos
 {
     public class RegistroVendasDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public DateOnly Data { get; set; }
         public double Quantidade { get; set; }
         public StatusVenda Status { get; set; }
-        public int VendedorId { get; set; }
+        public long VendedorId { get; set; }
         public List<Vendedor> Vendedores { get; set; }
         public List<Produto?> Produtos { get; set; }
     }
