+++ ProjetoMVC/Models/Dtos/VendedorDtos.cs
@@ -4,11 +4,11 @@ namespace ProjetoMVC.Models.Dtos
 {
     public class VendedorDtos
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Nome { get; set; }
         public string Email { get; set; }
         public double? SalarioBase { get; set; }
-        public int DepartamentoId { get; set; }
+        public long DepartamentoId { get; set; }
     }
 }
